fix-file-structure
