+++ api/teams/teams.type.ts
@@ -14,3 +14,9 @@ export type GetTeamsDto = {
   teams: TTeam[]
   total: number
 }
+export type TeamCreatePayload = {
+  name: string
+  description?: string
+  member_ids: string[]
+  poc_id: string | null
+}
